reorganize components into folders
